backend/requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
pdfplumber
python-docx
python-multipart
numpy==1.26.4